chore(outputs.cloud_pubsub): Merge topic stub into test code
